Move create workers to run command
